fix(web): fully restore composer container/view split after #72
#72 renamed composer-view into composer.tsx and deleted the container
adapters. The first restore left a half-state where composer-container
imported a missing view. Restore the pre-#72 composer layout and point
workspace-detail back at the container entrypoints.
