+++ apps/web/src/features/workspace-detail/workspace-detail-page.tsx
@@ -21,8 +21,8 @@ import { Spinner } from '~/components/ui/spinner'
 import { toastManager } from '~/components/ui/toast'
 import { runtimeComposerUsesCollapsedInput } from '~/features/agent-runtime/use-runtime-catalog'
 import { describeChatExecutionError } from '~/features/chat/commands/chat-execution-errors'
-import type { DraftChatComposerSubmitOptions } from '~/features/chat/composer/draft-chat-composer'
-import { DraftChatComposer } from '~/features/chat/composer/draft-chat-composer'
+import type { DraftChatComposerSubmitOptions } from '~/features/chat/composer/containers/draft-chat-composer-container'
+import { DraftChatComposer } from '~/features/chat/composer/containers/draft-chat-composer-container'
 import type { ChatContextPart } from '~/features/chat/context/chat-context-parts'
 import { readRunRuntimeSettingsPatch } from '~/features/chat/runtime/runtime-settings-presenter'
 import { startOptimisticChatResponse } from '~/features/chat/session/optimistic-chat-turn'
